+++ test/test-meddleware.js
@@ -174,6 +174,27 @@ test('module', function (t) {
         t.end();
     });
 
+    test('test local resolve edge case', function (t) {
+        var meddleware = require('../index');
+        var express = require('express');
+
+        var app = express();
+
+        var config = {
+            example: {
+                module: './rq.js'
+            }
+        };
+
+        try {
+            app.use(meddleware(config));
+        } catch (e) {
+            t.fail(e.message, 'error', e.message);
+        }
+
+        t.end();
+    });
+
     t.end();
 
 });
